backend/requirements.txt
fastapi
pandas
numpy
sqlalchemy
psycopg2-binary
scikit-learn
xgboost
joblib
streamlit
python-dotenv
uvicorn
uvicorn[standard]
requests
prophet  
prometheus_client
